fix calc for 32-bit architectures
Fix compiler warning for `custom/u_pfe.c`.

Update the `ARCH_CFLAGS` make symbol comments in `Makefile.config`.

Add `ARCH_CFLAGS` to `COMMON_CFLAGS` in `Makefile.config`.

Add "RHEL (Linux) Address Sanitizer (ASAN) for i686" commented out
section to `Makefile.local`.

Improve comments in `Makefile.local` related to "Address Sanitizer (ASAN)".

Add `junk4200` to `.gitignore`.

Use `-${Q}` instead of `-@` while build hsrc files in `Makefile`.

Fix `pseudo_seed()` in `seed.c` to handle 32-bit architectures.

Fixed `reopenid()` in `file.c` to handle 32-bit architectures and
to avoid "glibc detected an invalid stdio handle" using `freopen(3)`.

Fixed bug in `c_power()` in `comfunc.c` that caused 32-bit
architectures to dump core.

Fixed bug in `zrandom()` in `zrandom.c` where results from a
Blum-Blum-Shub operation were incorrectly copied into a ZVALUE
for 32-bit architectures.

Fixed comment typo in `cal/regress.cal` before test 5305.

+++ custom/u_pfe.c
@@ -1712,7 +1712,8 @@ u_pfe_pread(char *UNUSED(name), int count, VALUE **vals)
         if (FD_ISSET(out, &rcs)) {
             r = read(out, &ob, pfe_pfe_SIZE_BUFFER);
             if (r < 0) {
-                math_error("%s: " __FILE__ ": %d: (out, %lu) read: %s", custname, __LINE__, strlen(o), strerror(errno));
+                math_error("%s: " __FILE__ ": %d: (out, %lu) read: %s",
+                           custname, __LINE__, (unsigned long)strlen(o), strerror(errno));
             }
             if (r) {
                 ob[r] = '\0';
@@ -1728,7 +1729,8 @@ u_pfe_pread(char *UNUSED(name), int count, VALUE **vals)
         if (FD_ISSET(err, &rcs)) {
             r = read(err, &eb, pfe_pfe_SIZE_BUFFER);
             if (r < 0) {
-                math_error("%s: " __FILE__ ": %d: (err, %lu) read: %s", custname, __LINE__, strlen(e), strerror(errno));
+                math_error("%s: " __FILE__ ": %d: (err, %lu) read: %s",
+                           custname, __LINE__,(unsigned long)strlen(e), strerror(errno));
             }
             if (r) {
                 eb[r] = '\0';
